Don't report iterator conversion for mutating container member functions
Some mutating member functions in Qt6 collections now accept
const_iterators (contrary to Qt5) as arguments and are meant
to be callable with both const and non-const iterators. This
patch makes clazy skip reporting these in the strict-iterators
check.

Also adds a test qt6-strict-iterators which exposes the
issue which this patch fixes.